+++ packages/gestalt/src/Masonry/__integration__/handleItemUpdates.ghost.js
@@ -8,7 +8,7 @@ describe('Masonry > handle item updates', () => {
   it('should correctly update grid item heights', async () => {
     ghost.close();
 
-    await ghost.open('http://localhost:3000/Masonry?virtualize=1', {
+    await ghost.open('http://localhost:3001/Masonry?virtualize=1', {
       viewportSize: {
         width: 800,
         height: 800,