+++ packages/gestalt/src/Masonry/__integration__/renderHeight.ghost.js
@@ -7,7 +7,7 @@ import selectors from './lib/selectors';
 describe('Masonry > Render Height', () => {
   it('Items can be positioned under the grid', async () => {
     ghost.close();
-    await ghost.open('http://localhost:3000/Masonry?finiteLength=1');
+    await ghost.open('http://localhost:3001/Masonry?finiteLength=1');
 
     const gridItems = await ghost.findElements(selectors.gridItem);
     assert.ok(gridItems.length > 0);