+++ packages/gestalt/src/Masonry/__integration__/virtualBoundsVisibilityWithScrollContainer.ghost.js
@@ -15,7 +15,7 @@ describe('Masonry > virtual bounds visibility /w scroll container', () => {
     // First load the page with javascript disabled to get the item position
     const VIRTUALIZED_TOP = 800;
     await ghost.open(
-      `http://localhost:3000/Masonry?scrollContainer=1&offsetTop=${VIRTUALIZED_TOP}`,
+      `http://localhost:3001/Masonry?scrollContainer=1&offsetTop=${VIRTUALIZED_TOP}`,
       {
         viewportSize: {
           width: 800,